requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
astroid==3.0.2
attrs==23.2.0
blinker==1.9.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
clr-loader==0.2.6
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
dill==0.3.7
distro==1.9.0
EasyProcess==1.1
entrypoint2==1.1
Flask==3.1.0
fonttools==4.43.1
GPUtil==1.4.0
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.5
load-dotenv==0.1.0
MarkupSafe==3.0.2
matplotlib==3.8.0
mccabe==0.7.0
MouseInfo==0.1.3
mss==9.0.1
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.33
mysql-connector-python-rf==2.2.2
mysqlclient==2.2.0
numpy==1.26.0
openai==1.9.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
Pillow==10.0.0
platformdirs==4.1.0
pluggy==1.5.0
protobuf==3.20.3
psutil==5.9.5
py-cpuinfo==9.0.0
py-zabbix==1.1.7
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydocstyle==6.3.0
PyGetWindow==0.0.9
pylint==3.0.3
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyodbc==5.0.1
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pythonnet==3.0.2
pytweening==1.0.7
pytz==2023.3.post1
pyuac==0.0.3
pyxll==5.7.1
PyYAML==6.0.1
requests==2.31.0
selenium==4.16.0
six==1.16.0
SNAKES==0.9.32
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
tee==0.0.3
tomlkit==0.12.3
tqdm==4.66.1
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
webdriver-manager==4.0.1
Werkzeug==3.1.3
wsproto==1.2.0
XlsxWriter==3.1.9
